test(cypress): extract tests for verify-options
