+++ packages/schematics/src/cypress/index.spec.ts
@@ -47,35 +47,7 @@ describe('@co-it/schematics:cypress', () => {
       );
     });
 
-    it('should throw an error if app name is invalid', () => {
-      const parameters = {
-        ...defaultParameters,
-        app: 'not-existing'
-      };
-
-      expect(() =>
-        runner.runSchematic('cypress', parameters, treeBefore)
-      ).toThrowError(
-        'Invalid app name provided. ' +
-          'Please provide a name of an existing Angular app as configured in angular.json.'
-      );
-    });
-
     describe('with --overwrite=false', () => {
-      it('should throw an error if e2e project is found for specified app', () => {
-        const parameters = {
-          ...defaultParameters,
-          app: 'my-app'
-        };
-
-        expect(() =>
-          runner.runSchematic('cypress', parameters, treeBefore)
-        ).toThrowError(
-          'Existing project named "my-app-e2e" was found. ' +
-            'Please set --overwrite to true to overwrite existing project.'
-        );
-      });
-
       it('should throw an error if e2e project is found for default project', () => {
         const parameters = defaultParameters;
 
@@ -84,39 +56,6 @@ describe('@co-it/schematics:cypress', () => {
         ).toThrowError(
           'Existing project named "my-app-e2e" was found. ' +
             'Please set --overwrite to true to overwrite existing project.'
-        );
-      });
-    });
-
-    describe('with --overwrite=true', () => {
-      it('should throw an error if e2e project to be overwritten is not found', () => {
-        const parameters = {
-          ...defaultParameters,
-          app: 'app-without-e2e-tests',
-          overwrite: true
-        };
-
-        expect(() =>
-          runner.runSchematic('cypress', parameters, treeBefore)
-        ).toThrowError(
-          'No e2e project named "app-without-e2e-tests-e2e" was found. ' +
-            'If you want to create a new e2e project, please set --overwrite to false '
-        );
-      });
-
-      it('should throw an error if --folder is set', () => {
-        const parameters = {
-          ...defaultParameters,
-          app: 'my-app',
-          overwrite: true,
-          folder: 'some-folder'
-        };
-
-        expect(() =>
-          runner.runSchematic('cypress', parameters, treeBefore)
-        ).toThrowError(
-          'Parameters --overwrite and --folder are mutually exclusive. ' +
-            'Original root folder is used when overwriting an existing e2e project.'
         );
       });
     });
